refactor: use DI instead of config dir
