--- config/package.json
@@ -1,7 +1,0 @@
-{
-  "name": "@senken/config",
-  "main": "out/index.js",
-  "scripts": {
-    "build": "tsc"
-  }
-}
