+++ sql-extraction/ts/src/index.ts
@@ -1,6 +1,6 @@
-import type { SqlNode } from "@senken/config";
 import * as ts from "typescript";
 import * as v from "valibot";
+import type { SqlNode } from "../../../vsce/src/interface";
 
 export const customRawSqlQueryTsSchema = v.array(
   v.object({
