+++ sql-extraction/ts/tsconfig.json
@@ -1,7 +1,6 @@
 {
   "compilerOptions": {
     "outDir": "out",
-    "rootDir": "src",
     "esModuleInterop": false,
     "declaration": true
   },
