+++ vsce/src/interface.ts
